Moved views into base.py
