+++ corehq/apps/domain/urls.py
@@ -54,9 +54,9 @@ from corehq.apps.domain.views import (
     TransferDomainView,
     WireInvoiceView,
     calculated_properties,
-    select,
     toggle_diff,
 )
+from corehq.apps.domain.views.base import select
 from corehq.apps.domain.views.exchange import ExchangeSnapshotsView, set_published_snapshot,
 from corehq.apps.domain.views.fixtures import CalendarFixtureConfigView, LocationFixtureConfigView
 from corehq.apps.domain.views.pro_bono import ProBonoView
